+++ src/main/java/com/chicmic/trainingModule/Dto/rating/Rating_PPT.java
@@ -1,5 +1,4 @@
 package com.chicmic.trainingModule.Dto.rating;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -8,12 +7,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Rating_PPT implements Rating{
-    private Float communicationRating;
-    private Float technicalRating;
-    private Float presentationRating;
+    private Double communicationRating;
+    private Double technicalRating;
+    private Double presentationRating;
     private String courseId;
-    public Float computeOverallRating(){
-        float total = communicationRating + technicalRating + presentationRating;
+    public double computeOverallRating(){
+        double total = communicationRating + technicalRating + presentationRating;
         return total/3;
     }
 }
