+++ src/main/java/com/chicmic/trainingModule/Dto/rating/Rating_TEST.java
@@ -1,6 +1,5 @@
 package com.chicmic.trainingModule.Dto.rating;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -10,11 +9,11 @@ import lombok.Setter;
 @Setter
 public class Rating_TEST implements Rating{
     private String testId;
-    private Float theoreticalRating;
-    private Float codingRating;
-    private Float communicationRating;
-    public Float computeOverallRating(){
-        float total = communicationRating + theoreticalRating + codingRating;
+    private Double theoreticalRating;
+    private Double codingRating;
+    private Double communicationRating;
+    public double computeOverallRating(){
+        double total = communicationRating + theoreticalRating + codingRating;
         return total/3;
     }
 }
